export method to add class to svg
https://coveord.atlassian.net/browse/JSUI-1505
