+++ src/ui/Facet/FacetSettings.ts
@@ -12,6 +12,7 @@ import { PopupUtils, HorizontalAlignment, VerticalAlignment } from '../../utils/
 import * as _ from 'underscore';
 import 'styling/_FacetSettings';
 import { SVGIcons } from '../../utils/SVGIcons';
+import { SVGDom } from '../../utils/SVGDom';
 
 export interface IFacetSettingsKlass {
   new (sorts: string[], facet: Facet): FacetSettings;
@@ -64,7 +65,7 @@ export class FacetSettings extends FacetSort {
 
     this.settingsIcon = $$('span', { className: 'coveo-facet-settings-more' }).el;
     this.settingsIcon.innerHTML = SVGIcons.more;
-    SVGIcons.addClassToSVGInContainer(this.settingsIcon, 'coveo-facet-settings-more-svg');
+    SVGDom.addClassToSVGInContainer(this.settingsIcon, 'coveo-facet-settings-more-svg');
     this.settingsButton.appendChild(this.settingsIcon);
 
     this.settingsPopup = $$('div', { className: 'coveo-facet-settings-popup' }).el;
